refactor: introduce JibService abstraction layer for build
